Added comments and new functions
Added comments to each of the current functions.
Defined a function (select-numbers) that will take in a list, then get a
random value from that list. It will then remove that value from the
givin list, and then add that value to the list n. The function will
then check if the list n is equal to 6, meaning there are 6 numbers in
the list. If this is so then the function will end, if not then it will
repeat with the modified list as the input.