Evaluating correctly part of the time- rpn
For the past day I have been trying to get the rpn calculations that
equate to the target number to be stored and displayed for the user. I
currently have it working by checking everytime the calculate function
tries to evaluate an expression. I have also included extra features to
help lower the amount of expressions the function checks along with some
error handling when a small number tries to take away a bigger number.